requirements.txt
requests
websockets
ccxt
numpy
pandas
tensorflow
torch
scikit-learn
matplotlib
rich
kivy
flask
python-dotenv
shap  # New addition for SHAP
transformers
kivy-garden-matplotlib
vaderSentiment
newsapi-python
scipy
torch-fidelity
bayes_opt
pyjwt
cryptography
spektral
joblib
